styled color color picker form component
