--- src/styles/ColorPickerFormStyles.js
@@ -1,18 +1,0 @@
-const styles = {
-  picker: {
-    width: "100% !important",
-    marginTop: "2rem",
-  },
-  addColor: {
-    width: "100%",
-    padding: "1rem !important",
-    marginTop: "1rem !important",
-    fontSize: "1rem !important",
-  },
-  colorNameInput: {
-    width: "100%",
-    height: "70px",
-  },
-};
-
-export default styles;
